requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.40.0
chromadb==0.5.23
pydantic==2.9.2
python-dotenv==1.0.1
httpx>=0.27.0
